Refactoring UnityVersionResolver.cs
