+++ UnityEngineAnalyzer.CLI/Program.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using UnityEngineAnalyzer.CLI.Reporting;
 
@@ -11,7 +9,6 @@ namespace UnityEngineAnalyzer.CLI
     public class Program
     {
         private static readonly Dictionary<string, Type> AvailableExporters = new Dictionary<string, Type>();
-        private const UnityVersion DEFAULT_UNITY_VERSION = UnityVersion.LATEST;
 
         static Program()
         {
@@ -32,7 +29,8 @@ namespace UnityEngineAnalyzer.CLI
                     return;
                 }
 
-                options.Version = DefineUnityVersion(options);
+                var unityVersionResolver = new UnityVersionResolver();
+                options.Version = unityVersionResolver.ResolveVersion(options);
 
                 var startTime = DateTime.Now;
 
@@ -95,45 +93,6 @@ namespace UnityEngineAnalyzer.CLI
         }
 
         //TODO SET TO OWN CLASS
-        private static UnityVersion DefineUnityVersion(Options options)
-        {
-            if (options.Version != UnityVersion.NONE)
-            {
-                return options.Version;
-            }
 
-            //THIS ONLY WORKS ON UNITY >= 5, before that ProjectVersion.txt did not exists
-            var projectPath = new FileInfo(options.ProjectFile).Directory;
-            var projectVersionFile = new FileInfo(projectPath + "/ProjectSettings/ProjectVersion.txt");
-            if (projectVersionFile.Exists)
-            {
-                var projectVersionString = File.ReadAllText(projectVersionFile.FullName);
-                return TryParseUnityVersion(projectVersionString);
-            }
-
-            return DEFAULT_UNITY_VERSION;
-        }
-
-        //TODO UNIT TESTS
-        private static UnityVersion TryParseUnityVersion(string version)
-        {
-            string editorText = "m_EditorVersion: ";
-            var match = Regex.Match(version, editorText + "[0-9.a-z]*");
-
-            string src = match.Value.Substring(editorText.Length);
-            src = src.Replace('.', '_');
-            src = src.Substring(0, src.IndexOf('_') + 2);
-
-            var unityVersions = Enum.GetValues(typeof(UnityVersion)).Cast<UnityVersion>();
-            foreach (var unityVersion in unityVersions)
-            {
-                if (Enum.GetName(typeof(UnityVersion), unityVersion).Contains(src))
-                {
-                    return unityVersion;
-                }
-            }
-
-            return DEFAULT_UNITY_VERSION;
-        }
     }
 }
