+++ UnityEngineAnalyzer.CLI/UnityEngineAnalyzer.CLI.csproj
@@ -118,6 +118,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Reporting\StandardOutputAnalyzerReporter.cs" />
     <Compile Include="SolutionAnalyzer.cs" />
+    <Compile Include="UnityVersionResolver.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
